stop overriding currently attached xdp program and fix detach
Detach was not working when an interface supported native mode
and a xdp program was attached in generic mode.

Detach is internaly implemented by attaching with program file
descriptor -1, this change make it more clear and obvious why it
was not working (it's because it always tried to detach native mode xdp
programs, letting generic ones hanging).